requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-better-auth[fastapi]==0.1.2
psycopg2-binary==2.9.9
pydantic-settings==2.4.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
bcrypt==4.1.2